refactor: move bip39 to utils folder
