+++ apps/mobile/components/SSKeyboardWordSelector.tsx
@@ -13,11 +13,11 @@ import {
   ViewStyle
 } from 'react-native'
 
-import { getWordList } from '@/api/bip39'
 import useKeyboardHeight from '@/hooks/useKeyboardHeight'
 import usePrevious from '@/hooks/usePrevious'
 import { t } from '@/locales'
 import { Colors, Sizes } from '@/styles'
+import { getWordList } from '@/utils/bip39'
 
 type WordInfo = {
   index: number
