+++ apps/mobile/utils/bip39.ts
